+++ packages/cli/test/util/dev/validate.test.ts
